Refactor and enhance review command
Move `ReviewFeedbackItem` type to `options` for better organization and reuse. Add a check in `factory` to log a message and exit if no changes are detected, improving user feedback. These changes streamline the code and enhance the user experience by providing clearer information when no changes are present.
